Remove producing of summarymap config from the config model.
The search backend is no longer using this config.
