+++ config-model/src/main/java/com/yahoo/vespa/model/content/ContentSearchCluster.java
@@ -245,7 +245,7 @@ public class ContentSearchCluster extends AbstractConfigProducer<SearchCluster> 
                 throw new IllegalArgumentException("Schema '" + schemaDefinitionXMLHandler.getName() + "' referenced in " +
                                                    this + " does not exist");
 
-            sc.add(new SchemaInfo(schema, deployState.rankProfileRegistry(), null, null));
+            sc.add(new SchemaInfo(schema, deployState.rankProfileRegistry(), null));
         }
     }
