--- config-model/src/test/derived/advanced/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "location_zcurve"
-override[].command "attribute"
-override[].arguments "location_zcurve"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
