--- config-model/src/test/derived/array_of_struct_attribute/summarymap.cfg
@@ -1,16 +1,0 @@
-defaultoutputclass -1
-override[].field "new_elem_array"
-override[].command "attributecombiner"
-override[].arguments "elem_array"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "elem_array"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
