--- config-model/src/test/derived/attributeprefetch/summarymap.cfg
@@ -1,28 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "singlebyte"
-override[].command "attribute"
-override[].arguments "singlebyte"
-override[].field "singleint"
-override[].command "attribute"
-override[].arguments "singleint"
-override[].field "singlelong"
-override[].command "attribute"
-override[].arguments "singlelong"
-override[].field "singlefloat"
-override[].command "attribute"
-override[].arguments "singlefloat"
-override[].field "singledouble"
-override[].command "attribute"
-override[].arguments "singledouble"
-override[].field "singlestring"
-override[].command "attribute"
-override[].arguments "singlestring"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
