--- config-model/src/test/derived/attributes/summarymap.cfg
@@ -1,55 +1,0 @@
-defaultoutputclass -1
-override[].field "a1"
-override[].command "attribute"
-override[].arguments "a1"
-override[].field "a2"
-override[].command "attribute"
-override[].arguments "a2"
-override[].field "a3"
-override[].command "attribute"
-override[].arguments "a3"
-override[].field "a5"
-override[].command "attribute"
-override[].arguments "a5"
-override[].field "a6"
-override[].command "attribute"
-override[].arguments "a6"
-override[].field "b1"
-override[].command "attribute"
-override[].arguments "b1"
-override[].field "b2"
-override[].command "attribute"
-override[].arguments "b2"
-override[].field "b3"
-override[].command "attribute"
-override[].arguments "b3"
-override[].field "b4"
-override[].command "attribute"
-override[].arguments "b4"
-override[].field "b5"
-override[].command "attribute"
-override[].arguments "b5"
-override[].field "b6"
-override[].command "attribute"
-override[].arguments "b6"
-override[].field "b7"
-override[].command "attribute"
-override[].arguments "b7"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "a9"
-override[].command "attribute"
-override[].arguments "a9"
-override[].field "a11"
-override[].command "attribute"
-override[].arguments "a11"
-override[].field "a12"
-override[].command "attribute"
-override[].arguments "a12"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
