--- config-model/src/test/derived/complex/summarymap.cfg
@@ -1,34 +1,0 @@
-defaultoutputclass -1
-override[].field "dyntitle"
-override[].command "dynamicteaser"
-override[].arguments "dyntitle"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "year_sub"
-override[].command "attribute"
-override[].arguments "year_sub"
-override[].field "prefixenabled"
-override[].command "attribute"
-override[].arguments "prefixenabled"
-override[].field "fleeting2"
-override[].command "attribute"
-override[].arguments "fleeting2"
-override[].field "foundat"
-override[].command "attribute"
-override[].arguments "foundat"
-override[].field "collapseby"
-override[].command "attribute"
-override[].arguments "collapseby"
-override[].field "ts"
-override[].command "attribute"
-override[].arguments "ts"
-override[].field "combineda"
-override[].command "attribute"
-override[].arguments "combineda"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
