--- config-model/src/test/derived/imported_position_field_summary/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "my_pos"
-override[].command "geopos"
-override[].arguments "my_pos_zcurve"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
