--- config-model/src/test/derived/imported_struct_fields/summarymap.cfg
@@ -1,28 +1,0 @@
-defaultoutputclass -1
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
-override[].field "my_elem_array"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "my_elem_map"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "my_str_int_map"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "elem_array_filtered"
-override[].command "matchedattributeelementsfilter"
-override[].arguments "my_elem_array"
-override[].field "elem_map_filtered"
-override[].command "matchedattributeelementsfilter"
-override[].arguments "my_elem_map"
-override[].field "str_int_map_filtered"
-override[].command "matchedattributeelementsfilter"
-override[].arguments "my_str_int_map"
