--- config-model/src/test/derived/importedfields/summarymap.cfg
@@ -1,28 +1,0 @@
-defaultoutputclass -1
-override[].field "my_int_field"
-override[].command "attribute"
-override[].arguments "my_int_field"
-override[].field "my_string_field"
-override[].command "attribute"
-override[].arguments "my_string_field"
-override[].field "my_int_array_field"
-override[].command "attribute"
-override[].arguments "my_int_array_field"
-override[].field "my_int_wset_field"
-override[].command "attribute"
-override[].arguments "my_int_wset_field"
-override[].field "my_ancient_int_field"
-override[].command "attribute"
-override[].arguments "my_ancient_int_field"
-override[].field "my_filtered_int_array_field"
-override[].command "matchedattributeelementsfilter"
-override[].arguments "my_int_array_field"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
