--- config-model/src/test/derived/inheritance/summarymap.cfg
@@ -1,19 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "onlygrandparent"
-override[].command "attribute"
-override[].arguments "onlygrandparent"
-override[].field "overridden"
-override[].command "attribute"
-override[].arguments "overridden"
-override[].field "onlymother"
-override[].command "attribute"
-override[].arguments "onlymother"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
