+++ config-model/src/test/derived/inheritfromparent/summary.cfg
@@ -1,0 +1,36 @@
+defaultsummaryid 1561972036
+usev8geopositions true
+classes[].id 1561972036
+classes[].name "default"
+classes[].omitsummaryfeatures false
+classes[].fields[].name "weight"
+classes[].fields[].type "float"
+classes[].fields[].command "attribute"
+classes[].fields[].source "weight"
+classes[].fields[].name "rankfeatures"
+classes[].fields[].type "featuredata"
+classes[].fields[].command "rankfeatures"
+classes[].fields[].source ""
+classes[].fields[].name "summaryfeatures"
+classes[].fields[].type "featuredata"
+classes[].fields[].command "summaryfeatures"
+classes[].fields[].source ""
+classes[].fields[].name "documentid"
+classes[].fields[].type "longstring"
+classes[].fields[].command "documentid"
+classes[].fields[].source ""
+classes[].id 571976254
+classes[].name "attributeprefetch"
+classes[].omitsummaryfeatures false
+classes[].fields[].name "weight"
+classes[].fields[].type "float"
+classes[].fields[].command "attribute"
+classes[].fields[].source "weight"
+classes[].fields[].name "rankfeatures"
+classes[].fields[].type "featuredata"
+classes[].fields[].command "rankfeatures"
+classes[].fields[].source ""
+classes[].fields[].name "summaryfeatures"
+classes[].fields[].type "featuredata"
+classes[].fields[].command "summaryfeatures"
+classes[].fields[].source ""
