--- config-model/src/test/derived/inheritfromparent/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "weight"
-override[].command "attribute"
-override[].arguments "weight"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
