--- config-model/src/test/derived/mail/summarymap.cfg
@@ -1,16 +1,0 @@
-defaultoutputclass -1
-override[].field "snippet"
-override[].command "dynamicteaser"
-override[].arguments "snippet"
-override[].field "date"
-override[].command "attribute"
-override[].arguments "date"
-override[].field "attachmentcount"
-override[].command "attribute"
-override[].arguments "attachmentcount"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
