--- config-model/src/test/derived/map_attribute/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "str_map"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
