--- config-model/src/test/derived/map_of_struct_attribute/summarymap.cfg
@@ -1,19 +1,0 @@
-defaultoutputclass -1
-override[].field "new_str_elem_map"
-override[].command "attributecombiner"
-override[].arguments "str_elem_map"
-override[].field "new_int_elem_map"
-override[].command "copy"
-override[].arguments "int_elem_map"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "str_elem_map"
-override[].command "attributecombiner"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
