--- config-model/src/test/derived/multiplesummaries/summarymap.cfg
@@ -1,58 +1,0 @@
-defaultoutputclass -1
-override[].field "a"
-override[].command "attribute"
-override[].arguments "a"
-override[].field "adynamic"
-override[].command "dynamicteaser"
-override[].arguments "adynamic"
-override[].field "d"
-override[].command "dynamicteaser"
-override[].arguments "d"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "c"
-override[].command "attribute"
-override[].arguments "c"
-override[].field "adynamic2"
-override[].command "dynamicteaser"
-override[].arguments "adynamic2"
-override[].field "sometags"
-override[].command "matchedelementsfilter"
-override[].arguments "mytags"
-override[].field "anothera"
-override[].command "attribute"
-override[].arguments "a"
-override[].field "anotdynamic"
-override[].command "attribute"
-override[].arguments "adynamic"
-override[].field "loc_position"
-override[].command "attribute"
-override[].arguments "loc_pos_zcurve"
-override[].field "anotbolded"
-override[].command "attribute"
-override[].arguments "a"
-override[].field "loc_pos_zcurve"
-override[].command "attribute"
-override[].arguments "loc_pos_zcurve"
-override[].field "abolded2"
-override[].command "dynamicteaser"
-override[].arguments "abolded2"
-override[].field "aboldeddynamic"
-override[].command "dynamicteaser"
-override[].arguments "aboldeddynamic"
-override[].field "loc_pos"
-override[].command "geopos"
-override[].arguments "loc_pos_zcurve"
-override[].field "abolded"
-override[].command "dynamicteaser"
-override[].arguments "abolded"
-override[].field "dynamice"
-override[].command "dynamicteaser"
-override[].arguments "dynamice"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
