--- config-model/src/test/derived/newrank/summarymap.cfg
@@ -1,52 +1,0 @@
-defaultoutputclass -1
-override[].field "bgndata"
-override[].command "dynamicteaser"
-override[].arguments "bgndata"
-override[].field "sales"
-override[].command "attribute"
-override[].arguments "sales"
-override[].field "pto"
-override[].command "attribute"
-override[].arguments "pto"
-override[].field "mid"
-override[].command "attribute"
-override[].arguments "mid"
-override[].field "ew"
-override[].command "dynamicteaser"
-override[].arguments "ew"
-override[].field "weight"
-override[].command "attribute"
-override[].arguments "weight"
-override[].field "song"
-override[].command "dynamicteaser"
-override[].arguments "song"
-override[].field "bgnpfrom"
-override[].command "attribute"
-override[].arguments "bgnpfrom"
-override[].field "newestedition"
-override[].command "attribute"
-override[].arguments "newestedition"
-override[].field "bgnpto"
-override[].command "dynamicteaser"
-override[].arguments "bgnpto"
-override[].field "year"
-override[].command "attribute"
-override[].arguments "year"
-override[].field "did"
-override[].command "attribute"
-override[].arguments "did"
-override[].field "scorekey"
-override[].command "attribute"
-override[].arguments "scorekey"
-override[].field "cbid"
-override[].command "attribute"
-override[].arguments "cbid"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
