--- config-model/src/test/derived/position_nosummary/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "pos_zcurve"
-override[].command "attribute"
-override[].arguments "pos_zcurve"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
