--- config-model/src/test/derived/position_summary/summarymap.cfg
@@ -1,16 +1,0 @@
-defaultoutputclass -1
-override[].field "pos"
-override[].command "geopos"
-override[].arguments "pos_zcurve"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "pos_zcurve"
-override[].command "attribute"
-override[].arguments "pos_zcurve"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
