--- config-model/src/test/derived/predicate_attribute/summarymap.cfg
@@ -1,10 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
