--- config-model/src/test/derived/rankingexpression/summarymap.cfg
@@ -1,37 +1,0 @@
-defaultoutputclass -1
-override[].field "year"
-override[].command "attribute"
-override[].arguments "year"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "foo1"
-override[].command "attribute"
-override[].arguments "foo1"
-override[].field "foo2"
-override[].command "attribute"
-override[].arguments "foo2"
-override[].field "foo3"
-override[].command "attribute"
-override[].arguments "foo3"
-override[].field "foo4"
-override[].command "attribute"
-override[].arguments "foo4"
-override[].field "bar1"
-override[].command "attribute"
-override[].arguments "bar1"
-override[].field "bar2"
-override[].command "attribute"
-override[].arguments "bar2"
-override[].field "bar3"
-override[].command "attribute"
-override[].arguments "bar3"
-override[].field "bar4"
-override[].command "attribute"
-override[].arguments "bar4"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
