--- config-model/src/test/derived/schemainheritance/summarymap.cfg
@@ -1,16 +1,0 @@
-defaultoutputclass -1
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "parent_field"
-override[].command "attribute"
-override[].arguments "parent_field"
-override[].field "child_field"
-override[].command "attribute"
-override[].arguments "child_field"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
