--- config-model/src/test/derived/streamingstruct/summarymap.cfg
@@ -1,16 +1,0 @@
-defaultoutputclass -1
-override[].field "snippet"
-override[].command "dynamicteaser"
-override[].arguments "snippet"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "g"
-override[].command "dynamicteaser"
-override[].arguments "g"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
