--- config-model/src/test/derived/twostreamingstructs/summarymap.cfg
@@ -1,13 +1,0 @@
-defaultoutputclass -1
-override[].field "snippet"
-override[].command "dynamicteaser"
-override[].arguments "snippet"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "g"
-override[].command "dynamicteaser"
-override[].arguments "g"
