--- config-model/src/test/derived/types/summarymap.cfg
@@ -1,34 +1,0 @@
-defaultoutputclass -1
-override[].field "abyte"
-override[].command "attribute"
-override[].arguments "abyte"
-override[].field "along"
-override[].command "attribute"
-override[].arguments "along"
-override[].field "abool"
-override[].command "attribute"
-override[].arguments "abool"
-override[].field "ashortfloat"
-override[].command "attribute"
-override[].arguments "ashortfloat"
-override[].field "tagfield"
-override[].command "attribute"
-override[].arguments "tagfield"
-override[].field "album1"
-override[].command "attribute"
-override[].arguments "album1"
-override[].field "rankfeatures"
-override[].command "rankfeatures"
-override[].arguments ""
-override[].field "summaryfeatures"
-override[].command "summaryfeatures"
-override[].arguments ""
-override[].field "other"
-override[].command "attribute"
-override[].arguments "other"
-override[].field "juletre"
-override[].command "attribute"
-override[].arguments "juletre"
-override[].field "documentid"
-override[].command "documentid"
-override[].arguments ""
